refactor(mctrl): split api endpoints in modules
